requirements.txt
## The following requirements were added by pip freeze:
astroid==2.15.0
attrs==22.2.0
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
Flask==2.2.3
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pylint==2.17.1
pymongo==4.3.3
pytest==7.2.2
tomlkit==0.11.6
Werkzeug==2.2.3
wrapt==1.15.0
gunicorn
python-dotenv
